fix AbstractSessionManager logging calls
Calling factory.getClass().getEnclosingClass().getName() does not
account for the possibility that the factory is NOT an enclosed class
resulting in a NullPointerException when calling getName().

Instead we should assume that it could be either an enclosed or
top-level class.

This adds a new method: private String getFactoryClassName(...) to
handle the null check and return the appropriate class name.